chore: adjust base on comments
