+++ plugin/client/base/client_test.go
@@ -22,12 +22,10 @@ import (
 	"fmt"
 	"io"
 	"net/http"
-	"reflect"
 	"testing"
 
 	"github.com/go-resty/resty/v2"
 	"github.com/jarcoal/httpmock"
-	metav1alpha1 "github.com/katanomi/pkg/apis/meta/v1alpha1"
 	. "github.com/onsi/gomega"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
@@ -169,84 +167,4 @@ func TestPluginClientClone(t *testing.T) {
 
 	g.Expect(clone).To(Equal(original))
 	g.Expect(fmt.Sprintf("%p", original)).NotTo(Equal(fmt.Sprintf("%p", clone)))
-}
-
-func TestGetGitRepoInfo(t *testing.T) {
-	tests := []struct {
-		name        string
-		gitAddress  string
-		wantHost    string
-		wantGitRepo metav1alpha1.GitRepo
-		wantErr     bool
-	}{
-		{
-			name:        "invalid git repo url",
-			gitAddress:  "http:// github.com/katanomi/pkg.git",
-			wantHost:    "",
-			wantGitRepo: metav1alpha1.GitRepo{},
-			wantErr:     true,
-		},
-		{
-			name:       "shuffix with .git",
-			gitAddress: "http://github.com/katanomi/pkg.git",
-			wantHost:   "http://github.com",
-			wantGitRepo: metav1alpha1.GitRepo{
-				Project:    "katanomi",
-				Repository: "pkg",
-			},
-			wantErr: false,
-		},
-		{
-			name:       "shuffix without .git",
-			gitAddress: "http://github.com/katanomi/pkg",
-			wantHost:   "http://github.com",
-			wantGitRepo: metav1alpha1.GitRepo{
-				Project:    "katanomi",
-				Repository: "pkg",
-			},
-			wantErr: false,
-		},
-		{
-			name:       "shuffix without /",
-			gitAddress: "http://github.com/katanomi/pkg/",
-			wantHost:   "http://github.com",
-			wantGitRepo: metav1alpha1.GitRepo{
-				Project:    "katanomi",
-				Repository: "pkg",
-			},
-			wantErr: false,
-		},
-		{
-			name:        "path too short",
-			gitAddress:  "http://github.com/katanomi/",
-			wantHost:    "",
-			wantGitRepo: metav1alpha1.GitRepo{},
-			wantErr:     true,
-		},
-		{
-			name:       "path with subgroup",
-			gitAddress: "http://github.com/katanomi/subgroup/pkg/",
-			wantHost:   "http://github.com",
-			wantGitRepo: metav1alpha1.GitRepo{
-				Project:    "katanomi/subgroup",
-				Repository: "pkg",
-			},
-			wantErr: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			gotHost, gotGitRepo, err := GetGitRepoInfo(tt.gitAddress)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("GetGitRepoInfo() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if gotHost != tt.wantHost {
-				t.Errorf("GetGitRepoInfo() gotHost = %v, want %v", gotHost, tt.wantHost)
-			}
-			if !reflect.DeepEqual(gotGitRepo, tt.wantGitRepo) {
-				t.Errorf("GetGitRepoInfo() gotGitRepo = %v, want %v", gotGitRepo, tt.wantGitRepo)
-			}
-		})
-	}
 }
